+++ yolort/ultralytics/__init__.py
@@ -1,1 +1,3 @@
+from .models.common import *
+from .models.yolo import Model, Detect
 from .helpers import YOLOv5, load_model as load
